Update type for restart delay.

+++ src/tests/simulation/src/cluster.rs
@@ -487,7 +487,7 @@ impl Cluster {
             // so that the node is expired and removed from the cluster
             if rand::thread_rng().gen_bool(0.1) {
                 // max_heartbeat_interval_secs = 60
-                t += Duration::from_secs(opts.restart_delay_secs);
+                t += Duration::from_secs(opts.restart_delay_secs as u64);
             }
             tokio::time::sleep(t).await;
             tracing::info!("restart {name}");
@@ -592,7 +592,7 @@ pub struct KillOpts {
     pub kill_frontend: bool,
     pub kill_compute: bool,
     pub kill_compactor: bool,
-    pub restart_delay_secs: i32,
+    pub restart_delay_secs: u32,
 }
 
 impl KillOpts {
